Create calculate tax
Write a program to calculate tax given the following conditions:
a. If income is less than or equal to 1,50,000 then no tax
b. If taxable income is 1,50,001 – 3,00,000 the charge 10% tax
c. If taxable income is 3,00,001 – 5,00,000 the charge 20% tax
d. If taxable income is above 5,00,001 then charge 30% tax
Sample Input:
Enter the income: 200000
Sample Output:
Tax= 20000